requirements.txt
google-api-python-client>=2.100.0,<3.0.0
google-auth>=2.23.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth-httplib2>=0.1.0,<1.0.0
pandas>=2.1.0,<3.0.0
matplotlib>=3.8.0,<4.0.0
reportlab>=4.4.0,<5.0.0
python-dateutil>=2.8.0,<3.0.0
